Telegram: move parse functions to platform module
These functions were developed specifically for Telegram, so they were
moved from the `util` module directly to the `telegram` platform module.
